Refactor: ♻︎ Refactor RankingView into separate components and screens
このコミットでは、RankingViewコンポーネントを二つの独立した部分に分割しました：

1. ランキングリストをレンダリングする役割を持つRankingListコンポーネントを作成し、`components`フォルダに移動しました。
2. RankingListコンポーネントを含むコンテナであるRankingViewスクリーンを`screens`フォルダに移動しました。

このようにすることで、コードの可読性と再利用性を向上します。
