+++ App.js
@@ -3,7 +3,7 @@ import { View, FlatList } from 'react-native';
 import { Button } from 'react-native-elements';
 import AlarmCard from './components/AlarmCard';
 import AlarmSettingOverlay from './components/AlarmSettingOverlay';
-import RankingView from './RankingView';
+import RankingView from './screens/RankingView';
 
 export default function App() {
   const [alarms, setAlarms] = useState([]);
