+++ screens/RankingView.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { View, Text, FlatList, StyleSheet } from 'react-native';
+import { View, Text, StyleSheet } from 'react-native';
+import RankingList from '../components/RankingList';
 
 const RankingView = () => {
 
@@ -29,29 +30,10 @@ const RankingView = () => {
 
   const newRankingData = makeUserRanking(userData, rankingData);
 
-  const switchBgColor = (id) => {
-    if (id === userData.id) {
-      return { backgroundColor: 'blue' };
-    }
-    else {
-      return { backgroundColor: 'white' };
-    }
-  }
-
   return (
     <View style={styles.container}>
       <Text style={styles.title}>Ranking</Text>
-      <FlatList 
-        data={newRankingData}
-        keyExtractor={item => item.id.toString()}
-        renderItem={({ item, index }) => (
-          <View style={[styles.listItem, switchBgColor(item.id)]}>
-            <Text style={styles.rank}>{index + 1}</Text>
-            <Text style={styles.name}>{item.name}</Text>
-            <Text style={styles.score}>{item.time}sec.</Text>
-          </View>
-        )}
-      />
+      <RankingList data={newRankingData} userData={userData} />
     </View>
   );
 };
@@ -66,26 +48,6 @@ const styles = StyleSheet.create({
     marginTop: 45,
     marginBottom: 16,
     textAlign: 'center',
-  },
-  listItem: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingVertical: 8,
-    borderBottomWidth: 1,
-    borderColor: '#ccc',
-  },
-  rank: {
-    fontSize: 16,
-    width: 30,
-  },
-  name: {
-    flex: 1,
-    fontSize: 16,
-  },
-  score: {
-    fontSize: 16,
-    width: 50,
   },
 });
